+++ spec/ruby/optional/box/fixtures/v1/box_spec_lib.rb
@@ -1,0 +1,3 @@
+module BoxSpecLib
+  VERSION = "1.0"
+end