+++ spec/ruby/optional/box/fixtures/v2/box_spec_lib.rb
@@ -1,0 +1,3 @@
+module BoxSpecLib
+  VERSION = "2.0"
+end